add tests to show lookupplan coherence
