+++ test/Chainweb/Data/Test/Backfill.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE NumericUnderscores #-}
 module Chainweb.Data.Test.Backfill
 ( tests
 ) where
@@ -6,11 +7,14 @@ module Chainweb.Data.Test.Backfill
 import Control.Arrow ((&&&))
 
 import Data.Bifunctor
+import Data.Function
+import Data.List (sortOn)
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as M
 import Data.Maybe (mapMaybe)
 
 import Chainweb.Api.ChainId (ChainId(..))
+import ChainwebData.Backfill
 import ChainwebData.Types
 
 import Test.Tasty
@@ -19,37 +23,17 @@ import Test.Tasty.HUnit
 
 tests :: TestTree
 tests = testGroup "Backfill tests"
-    [ backfillUnitTests
+    [ lookupPlanTests
     ]
 
 
-backfillUnitTests :: TestTree
-backfillUnitTests = testGroup "Backfill unit tests"
-    [ testCase "noop" (return ())
+lookupPlanTests :: TestTree
+lookupPlanTests = testGroup "Lookup plan unit tests"
+    [ testCase "New/Old Lookup plans should be isomorphic" $ do
+      let a = sortOn (\(t,_,_) -> t) $ oldLookupPlan testData
+          b = sortOn (\(t,_,_) -> t) $ lookupPlan testData
+      a @=? b
+
     ]
-
--- -------------------------------------------------------------------- --
--- Utils
-
--- | Based on some initial minimum heights per chain, form a lazy list of block
--- ranges that need to be looked up.
---
--- TODO: Parametrize by chaingraph history, genesis for each chain id
---
-oldLookupPlan :: Map ChainId Int -> [(ChainId, Low, High)]
-oldLookupPlan mins = concatMap (\pair -> mapMaybe (g pair) asList) ranges
   where
-    maxi :: Int
-    maxi = max 0 $ maximum (M.elems mins) - 1
-
-    asList :: [(ChainId, High)]
-    asList = map (second (\n -> High . max 0 $ n - 1)) $ M.toList mins
-
-    ranges :: [(Low, High)]
-    ranges = map (Low . last &&& High . head) $ groupsOf 100 [maxi, maxi-1 .. 0]
-
-    g :: (Low, High) -> (ChainId, High) -> Maybe (ChainId, Low, High)
-    g (l@(Low l'), u) (cid, mx@(High mx'))
-      | u > mx && l' <= mx' = Just (cid, l, mx)
-      | u <= mx = Just (cid, l, u)
-      | otherwise = Nothing
+    testData = M.fromList [ (ChainId x, 1_000) | x <- [0..9] ]
